camera-related: Use short instead of float and simplify camera angle normalization using modulo operation

+++ source/headers.h
@@ -15,5 +15,5 @@ void MoveCharacter3D(Vector3 *position, Camera3D *camera, int isNights);
 void JumpCharacter3D(Vector3 *position, Camera3D *camera);
 
 // camera
-void RotateCamera(Camera3D *camera, Vector3 *characterPosition, float *cameraAngle);
-void SwitchCameraTo2dot5D(Camera3D *camera, int isNights, float *cameraAngle);
+void RotateCamera(Camera3D *camera, Vector3 *characterPosition, short *cameraAngle);
+void SwitchCameraTo2dot5D(Camera3D *camera, int isNights, short *cameraAngle);
